backend/requirements.txt
# AquaMetric Backend Dependencies
# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Scientific Computing (SwimBIT Algorithm)
numpy>=1.26.0
scipy>=1.11.0
pandas>=2.1.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Task Queue (async processing)
celery>=5.3.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
httpx>=0.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Protobuf for efficient data transfer
protobuf>=4.25.0
